requirements.txt
google-generativeai
python-dotenv
beautifulsoup4
playwright
requests
